This project is a three-page application allowing users to explore a variety of coffees, add them to their cart, and seamlessly make payments.
Welcome to the Thalassophile Coffee , a simple yet delightful application designed to bring the joy of coffee customization to users. This project is a three-page application allowing users to explore a variety of coffees, add them to their cart, and seamlessly make payments. The use of Riverpod provider ensures smooth state management, and future enhancements can easily introduce internet-based features for a more connected experience. My aim is to create a visual, simple design.